Move timeouts to shared
