+++ shared/src/types/control_flow.rs
@@ -1,0 +1,3 @@
+//! Control flow utilities.
+
+pub mod timeouts;
